requirements.txt
streamlit
youtube_transcript_api
pytube
PyMuPDF
OpenAI
streamlit-authenticator
stripe
google-cloud-firestore
python-dotenv